// web/src/api/generated/models/course-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NovaCloudEdu API
 * NovaCloudEdu Smart Education Platform Backend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 课程信息响应
 */
export interface CourseResponse {
    /**
     * 课程ID
     */
    'id'?: number;
    /**
     * 课程标题
     */
    'title'?: string;
    /**
     * 课程副标题
     */
    'subtitle'?: string;
    /**
     * 课程描述
     */
    'description'?: string;
    /**
     * 封面图片URL
     */
    'coverImage'?: string;
    /**
     * 课程价格
     */
    'price'?: number;
    /**
     * 课程类型：0-公开课，1-付费课，2-会员课
     */
    'courseType'?: number;
    /**
     * 课程类型描述
     */
    'courseTypeDesc'?: string;
    /**
     * 难度等级：1-入门，2-初级，3-中级，4-高级，5-专家
     */
    'difficulty'?: number;
    /**
     * 难度等级描述
     */
    'difficultyDesc'?: string;
    /**
     * 状态：0-未发布，1-已发布，2-已下架
     */
    'status'?: number;
    /**
     * 状态描述
     */
    'statusDesc'?: string;
    /**
     * 讲师ID
     */
    'teacherId'?: number;
    /**
     * 总时长(分钟)
     */
    'totalDuration'?: number;
    /**
     * 总章节数
     */
    'totalChapters'?: number;
    /**
     * 总小节数
     */
    'totalSections'?: number;
    /**
     * 学习人数
     */
    'studentCount'?: number;
    /**
     * 评分
     */
    'ratingScore'?: number;
    /**
     * 标签列表
     */
    'tags'?: Array<string>;
    /**
     * 创建时间
     */
    'createTime'?: string;
    /**
     * 更新时间
     */
    'updateTime'?: string;
}